+++ src/helpers/configs.js
@@ -1,0 +1,5 @@
+const endpoint = 'localhost:5050'
+
+export {
+  endpoint
+}
